Avoid the need for cloning translation objects
Currently, we rely on clone() when passing translation objects around in
map_range(), which means that translation objects need to be clonable.

However, in the future the implementation of the Translation trait could be
provided by the external caller, and requiring the object to implement
the Clone trait or clone()'ing the object freely as we do may interfere
with its potentially stateful nature in other contexts.

So reorganize the code and add some lifetime annotations so that the
translation object is borrowed only a single time.