Mark page tables as Send
PageTableWithLevel is not automatically Send because of the contained
NonNull.

Explicitly implement the trait since it is in face safe (which then
means all the related types up to and including Mapping are
automatically Send).

This allows them to be accessed from multiple threads with appropriate
synchronization.

We do not, of course, mark any of these types as Sync.