Implement walk_range for traversing page tables
Unlike modify_range(), this uses a bound on the type F defining the
prototype of the callback function. This is necessary because walking
the page tables using an immutable reference to the Mapping object
requires a mutable closure, and this cannot be achieved using a dynamic
function type like PteUpdater.