Use statically typed callback in modify_range
PteUpdater is a dynamic function type, and so modify_range(), which
takes a reference to an instance of this type, can accept any argument
that implements the PteUpdater trait, even if the actual type is not
even known at compile time. This involves dynamic dispatch at runtime
(i.e., using vtables) which is more than we need in most cases.

So instead, let's use a reference to a generic type F, with a bound that
describes the expected prototype. This is more idiomatic for Rust, and
will result in a statically typed variant being instantiated by the
compiler, rather than relying on dynamic dispatch at runtime.

In order to retain support for cases where the callback is actually
dynamically typed, add the ?Sized bound to the type of F too so it can
be satisfied by references to PteUpdater as well.