Don't pass mutable table entry descriptors to modify_range() callers
When traversing the page table tree in modify_range(), we pass a mutable
reference to each encountered entry to a callback function, so that the
caller can make changes to the permission attributes. Traversing the
page tables in this manner cannot and should not permit the tree
structure of the page table hierarchy to be corrupted, given that it
could lead to memory leaks and bogus memory references (if the callback
converts a table entry into a block entry or vice versa)

So let's invoke the callback only for block and page entries, and double
check that the entry has not been converted into a table entry by the
callback routine.