Reject TABLE_OR_PAGE attribute in map_range()
Currently, map_range() accepts the TABLE_OR_PAGE attribute as a flag,
and happily sets it on what it assumes to be block mappings, even though
setting the bit results in the descriptors in question to be interpreted
as table descriptors, turning the output address into a dangling
pointer.

So reject this attribute.