Add mapping constraints to map_range() API
Add a constraints argument to the map_range() API and use it to
implement a NO_BLOCK_MAPPINGS constraint which ensures that all created
mappings are mapped down to pages.

Also add a NO_CONTIGUOUS_HINT flag for future use - we don't actually
ever set that bit so currently there is no need to test for it yet.