Return the subrange that violates break-before-make on BBM failure
Passing the entire range back to the caller on a break-before-make (BBM)
violation is not very helpful. We might add a violation type to the enum
later but let's at least return the subrange that actually triggered the
failure, which is expected to be far more useful to the caller.