Move TTBR preserve/restore out of the API
The fact that activating and deactivating a mapping keeps track of the
old TTBR value under the hood is problematic for concurrent use of the
API by multiple CPUs. So instead, make it the caller's job to keep track
of the TTBR, by returning the old value from activate(), and expecting
that old value to be provided when calling deactivate().

Now that the library no longer records the old TTBR value, and no longer
panicks on nested calls to activate()/deactivate() [which may occur from
different CPUs], replace the notion of 'active' with a counter, which is
incremented for each activation and decremented for each deactivation.

One consequence of these changes is that it is no longer possible to
implicitly deactivate a mapping when it is dropped, as the library does
not know which TTBR value to program in to the register. So instead,
make it mandatory to deactivate all activations of a mapping before it
goes out of scope.