Deal with memory ordering in page tables
Even though we model them as arbitrary arrays of usize in memory, live
page tables are actually referenced by the page table walker, which is a
separate observer in ARM ARM parlance, and this means that non-atomic
updates may result in the page table walker transiently observing bogus
values.

So instead, treat it as an array of AtomicUsize, so that we obtain the
following properties:

- no Copy or Clone traits, and so all manipulations of descriptors are
  guaranteed to operate on the descriptor directly, using atomic
  accesses

- all stores to descriptors are issued with release semantics, which
  guarantees that the page table walker will never observe stores in the
  wrong order (e.g., a table entry being created before the next-level
  table in question has been fully wiped [using ordinary stores])

Even though it does not implement the Clone trait, the Descriptor type
does implement a crate-private clone() method, for where it is needed
explicitly (there is a single case where a descriptor is copied into a
local variable to perform checks on it).