Drop dependency on zerocopy
Replace the zerocopy dependency with a simple reimplementation of
write_to(). This is needed because zerocopy requires the Immutable
trait, making it incompatible with AtomicUsize, which relies on interior
mutability for nebulous reasons.

Given that zerocopy is only needed for offline generation of page tables
in a very particular use case, we can drop it without any loss of
functionality.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>