Update modify_flags() to return an error rather than panic!()
Tighten modify_flags() to only permit modifications that are allowed by
the architecture when the descriptor is live. Instead of panicking when
this condition is violated, update the return type and return a Err() in
that case, and leave it to the caller to decide whether or not panicking
is a reasonable course of action.

On success, return a boolean that indicates whether or not the
descriptor was modified as a result. This may be used by the caller to
perform TLB invalidation.