Make the translation regime part of the Translation trait
In order to have access to the translation regime at the level where
individual descriptors are being manipulated, move its definition into
the translation type, and make it accessible via a trait method
'regime()'.

This is needed for fine grained TLB invalidation, where entries are only
invalidated in the TLB when modified. TLB invalidation takes the
exception level of the translation regime into account, which is why
this information is required.