Create separate file for descriptors, attributes and memory addresses
Before refactoring how descriptors are manipulated and passed to the
callback closures when modifying ranges, split off the related
definitions into a separate source file. This keeps things more
maintainable in general, but also permits us to keep the internals of a
descriptor and how it is modified/updated hidden from the rest of the
code.