Encapsulate descriptors when passing them via the callback routines
Currently, mutable descriptors are being passed to the callback routines
provided to the modify_range() API, which is problematic for a number of
reasons:

- the descriptor may be live, in which case BBM rules apply, and TLB
  maintenance should be carried out if/when the descriptor is modified,
  but keeping track of this based on a bare descriptor is difficult;

- the descriptor itself does not (and cannot) keep track of whether or
  not it is live;

- the descriptor does not know at which level it appears in the page
  table hierarchy, which makes some checks difficult to perform;

- BBM checks perform a dry run of the updates in order to minimize the
  risk hitting a BBM failure halfway through, and this currently
  requires an actual AtomicUsize in memory, even though the calculation
  may be done entirely in registers or in memory that does not require
  atomicity.

Address these concerns by encapsulating a mutable reference to the
underlying descriptor in a wrapper object that provides access to the
contents of the descriptor, and permits modifications. BBM checks will
be moved here, as well as TLB maintenance, in subsequent patches.