Move BBM checks into descriptor
Now that descriptors are being wrapped in objects that track the level
at which they appear, and whether or not they are live, the BBM checks
that reason in terms of attribute changes of the descriptors can be
moved in there as well, which is a more natural place for them.

But more importantly, it ensures that the mutable descriptors that are
presented to callers via the modify_range() API can only be modidied in
a BBM compliant manner if they are live, but can be modified or set more
freely when they are not.