Don't split block mappings if the output address and flags are the same
Currently, any attempt to create a mapping that intersects a block
mapping without covering it entirely will result in a split of that
block mapping. However, if the attributes of the newly created mapping
match the ones of the existing block mapping, creating the new entries
is redundant, making the split unnecessary as well.

So detect this case both in the mapping code and the BBM rules checker,
and simply retain the block mapping in this case. Also update the tests
accordingly.